authoriztion implemented

+++ src/api/notes/handler.js
@@ -8,7 +8,11 @@ class NotesHandler {
     this._validator.validateNotePayload(request.payload);
 
     const { title = 'untitled', body, tags } = request.payload;
-    const noteId = await this._service.addNote({ title, body, tags });
+    const { id: credentialId } = request.auth.credentials;
+
+    const noteId = await this._service.addNote({
+      title, body, tags, owner: credentialId,
+    });
     const response = h.response({
       status: 'success',
       message: 'Catatan berhasil ditambahkan',
@@ -20,8 +24,9 @@ class NotesHandler {
     return response;
   };
 
-  getNotesHandler = async () => {
-    const notes = await this._service.getNotes();
+  getNotesHandler = async (request) => {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: { notes },
@@ -30,6 +35,8 @@ class NotesHandler {
 
   getNoteByIdHandler = async (request) => {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
     const note = await this._service.getNoteById(id);
     return {
       status: 'success',
@@ -38,10 +45,13 @@ class NotesHandler {
   };
 
   putNoteByIdHandler = async (request, h) => {
-    const { id } = request.params;
-
     this._validator.validateNotePayload(request.payload);
-    this._service.editNoteById(id, request.payload);
+
+    const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
+    await this._service.editNoteById(id, request.payload);
 
     const response = h.response({
       status: 'success',
@@ -53,7 +63,9 @@ class NotesHandler {
 
   deleteNoteByIdHandler = async (request) => {
     const { id } = request.params;
-    this._service.deleteNoteById(id);
+    const { id: credentialId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
+    await this._service.deleteNoteById(id);
     return {
       status: 'success',
       message: 'Catatan berhasil dihapus',
